Modified thresholds in over/under-correction tests to compensate for PSF changes in a previous update
The test thresholds are somewhat arbitrary. As the PSF is now wider than when the thresholds
were determined, 3 tests failed. Visual inspection showed expected behaviour so we've raised
the thresholds.

However, we currently need a ridicilously high threshold for the RBV over-correction case.
The images show 4 isolated pixels with high values.
TODO: Can we test for what we know will happen?